+++ src/components/locale/cultures/bg-BG.js
@@ -44,14 +44,15 @@ Soho.Locale.addCulture('bg-BG', {
   }],
   // numbers/currencyFormats-numberSystem-latn/standard
   currencySign: 'лв',
-  currencyFormat: '#,##0.00 ¤',
+  currencyFormat: '### ¤',
   // numbers/symbols-numberSystem-latn
   numbers: {
     percentSign: '%',
-    percentFormat: '#,##0 %',
+    percentFormat: '### %',
     minusSign: '-',
     decimal: ',',
-    group: ' '
+    group: ' ',
+    groupSizes: [3, 3]
   },
   // Resx - Provided By Translation Team
   messages: {