+++ src/components/locale/cultures/en-GB.js
@@ -45,14 +45,15 @@ Soho.Locale.addCulture('en-GB', {
   }],
   // numbers/currencyFormats-numberSystem-latn/standard
   currencySign: '£',
-  currencyFormat: '¤#,##0.00',
+  currencyFormat: '¤###',
   // numbers/symbols-numberSystem-latn
   numbers: {
     percentSign: '%',
-    percentFormat: '#,##0 %',
+    percentFormat: '### %',
     minusSign: '-',
     decimal: '.',
-    group: ','
+    group: ',',
+    groupSizes: [3, 3]
   },
   // Resx - Provided By Translation Team
   messages: {