+++ src/components/locale/cultures/uk-UA.js
@@ -43,14 +43,15 @@ Soho.Locale.addCulture('uk-UA', {
   }],
   // numbers/currencyFormats-numberSystem-latn/standard
   currencySign: '₴',
-  currencyFormat: '#,##0.00 ¤',
+  currencyFormat: '### ¤',
   // numbers/symbols-numberSystem-latn
   numbers: {
     percentSign: '%',
-    percentFormat: '#,##0%',
+    percentFormat: '###%',
     minusSign: '-',
     decimal: ',',
-    group: ' '
+    group: ' ',
+    groupSizes: [3, 3]
   },
   // Resx - Provided By Translation Team
   messages: {